requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
channels>=4.0
channels-redis>=4.2
daphne>=4.0
redis>=5.0
numpy>=1.26
